Implement integration testing for geomad to 3 buckets. Refactor geomad and other commands to improve code quality and reusability.
