+++ ldn/aws_credentials.py
@@ -4,21 +4,20 @@
 
 import logging
 import os
-from contextlib import contextmanager
+from functools import partial
 
 import boto3
 import obstore
 from botocore.client import BaseClient
 from dep_tools.writers import write_to_s3
 
-from ldn.utils import AWS_REGION
+from ldn.utils import AWS_REGION, is_source_coop
 
 logger = logging.getLogger(__name__)
 
-
-_WRITE_KEY = "AWS_WRITE_ACCESS_KEY_ID"
-_WRITE_SECRET = "AWS_WRITE_SECRET_ACCESS_KEY"
-_WRITE_TOKEN = "AWS_WRITE_SESSION_TOKEN"
+_WRITE_KEY = "SOURCE_COOP_AWS_ACCESS_KEY_ID"
+_WRITE_SECRET = "SOURCE_COOP_AWS_SECRET_ACCESS_KEY"
+_WRITE_TOKEN = "SOURCE_COOP_AWS_SESSION_TOKEN"
 
 
 def get_write_session() -> boto3.Session:
@@ -31,80 +30,55 @@ def get_write_session() -> boto3.Session:
     key = os.environ.get(_WRITE_KEY)
     secret = os.environ.get(_WRITE_SECRET)
 
+    _is_source_coop = is_source_coop()
+    logger.info(f"is_source_coop={_is_source_coop}, key set={bool(key)}, secret set={bool(secret)}")
+
+    if not _is_source_coop:
+        logger.info("SOURCE_COOP_PUBLIC_URL or prefixes not set; skipping write credential setup.")
+        return boto3.Session(region_name=AWS_REGION)
+
     if not (key and secret):
         logger.info("AWS_WRITE_* env vars not set; falling back to default credential chain for writes.")
-        return boto3.Session()
+        return boto3.Session(region_name=AWS_REGION)
 
     logger.info("Using explicit write credentials from environment.")
+    logger.info(f"Write session key: {key[:8]}...")
     return boto3.Session(
         aws_access_key_id=key,
         aws_secret_access_key=secret,
         aws_session_token=os.environ.get(_WRITE_TOKEN),
-        region_name=AWS_REGION,  # Same for Source.Coop and all other buckets.
+        region_name=AWS_REGION,
     )
 
 
-def make_write_function(
-    session: boto3.Session,
-):
-    client: BaseClient = session.client("s3")
+def make_write_function(session: boto3.Session):
+    """Create a write function bound to the credentials of the given session.
 
-    def write_with_credentials(data, path, bucket, **kwargs):
-        return write_to_s3(data, path, bucket, client=client, **kwargs)
+    Args:
+        session: A boto3 session with the credentials to use for writing.
 
-    return write_with_credentials
-
-
-def get_write_client(session: boto3.Session) -> BaseClient:
-    return session.client("s3")
-
-
-def _session_credentials(session: boto3.Session) -> dict:
-    """Extract resolved credentials from a boto3 session as a plain dict."""
-    creds = session.get_credentials().get_frozen_credentials()
-    return {
-        "access_key_id": creds.access_key,
-        "secret_access_key": creds.secret_key,
-        "token": creds.token,  # None for long-lived creds
-        "region": session.region_name,
-    }
-
-
-# The env-var swap in write_credentials_as_env is not thread-safe. Since make_mosaics is synchronous/sequential
-# this is fine, but worth noting if it ever gets parallelised.
-@contextmanager
-def write_credentials_as_env(session: boto3.Session):
-    """Context manager that temporarily sets AWS env vars from a boto3 session.
-
-    Used for clients like MosaicBackend that build their own boto3 client
-    internally and can't accept an explicit client or session.
-    Restores the original environment on exit.
+    Returns:
+        A partial of :func:`write_to_s3` with the session's S3 client bound.
     """
-    creds = _session_credentials(session)
-
-    overrides = {
-        "AWS_ACCESS_KEY_ID": creds["access_key_id"],
-        "AWS_SECRET_ACCESS_KEY": creds["secret_access_key"],
-    }
-    if creds["token"]:
-        overrides["AWS_SESSION_TOKEN"] = creds["token"]
-    if creds["region"]:
-        overrides["AWS_DEFAULT_REGION"] = creds["region"]
-
-    originals = {k: os.environ.get(k) for k in overrides}
-    try:
-        os.environ.update(overrides)
-        yield
-    finally:
-        for k, v in originals.items():
-            if v is None:
-                os.environ.pop(k, None)
-            else:
-                os.environ[k] = v
+    client: BaseClient = session.client("s3")
+    creds = session.get_credentials().get_frozen_credentials()  # TODO: Remove
+    logger.info(f"Write function using key: {creds.access_key[:8]}...")  # TODO: Remove
+    return partial(write_to_s3, client=client)
 
 
+# Used for indexing
 def make_obstore_s3(bucket: str, session: boto3.Session) -> "obstore.store.S3Store":
     """Create an obstore S3Store using credentials from the given boto3 session."""
+
+    def _session_credentials(session: boto3.Session) -> dict:
+        """Extract resolved credentials from a boto3 session as a plain dict."""
+        creds = session.get_credentials().get_frozen_credentials()
+        return {
+            "access_key_id": creds.access_key,
+            "secret_access_key": creds.secret_key,
+            "token": creds.token,  # None for long-lived creds
+            "region": session.region_name,
+        }
 
     creds = _session_credentials(session)
     kwargs = {
