+++ ldn/cli_geomad.py
@@ -2,20 +2,15 @@ import logging
 import sys
 from typing import Literal
 
-import boto3
 import typer
 from dask.distributed import Client as DaskClient
 from dask.distributed import KilledWorker
-from dep_tools.aws import object_exists
 from dep_tools.exceptions import EmptyCollectionError
 from dep_tools.loaders import OdcLoader
 from dep_tools.searchers import PystacSearcher
-from dep_tools.stac_utils import StacCreator
-from dep_tools.writers import AwsDsCogWriter, AwsStacWriter
 from odc.stac import configure_s3_access
 from typing_extensions import Annotated
 
-from ldn.aws_credentials import get_write_client, get_write_session, make_write_function
 from ldn.geomad import (
     LANDSAT_BANDS,
     LANDSAT_OFFSET,
@@ -25,21 +20,23 @@ from ldn.geomad import (
     GeoMADProcessor,
     InsufficientScenesError,
 )
-from ldn.geomad import AwsStacTask as Task
+from ldn.geomad import (
+    AwsStacTask as Task,
+)
 from ldn.grids import get_gridspec
-from ldn.raster import PrefixedS3ItemPath
+from ldn.raster import build_pipeline_components
 from ldn.utils import (
-    AWS_REGION,
-    BUCKET,
     GEOMAD_DATASET_ID,
     GEOMAD_VERSION,
     LS7_YEAR_THRESHOLD,
     NON_PACIFIC_OWNER,
     PACIFIC_OWNER,
-    SENSOR,
-    SOURCE_COOP_PREFIX_GEOMAD,
-    SOURCE_COOP_PUBLIC_URL,
+    get_bucket,
+    get_full_path_prefix,
+    get_source_coop_config,
+    is_source_coop,
     owner_for_region,
+    parse_tile_id,
 )
 
 geomad_app = typer.Typer()
@@ -49,8 +46,8 @@ EXIT_OOM = 42  # KilledWorker — retry with more resources
 EXIT_SKIP = 43  # too few scenes / no items — expected, don't retry
 
 
-def count_scenes(
-    tile_id: str = typer.Option(..., help="Tile ID to count scenes for."),
+def _count_scenes(
+    tile_id_tuple: tuple[int, int] = typer.Option(..., help="Tile ID tuple to count scenes for."),
     region: Literal["pacific", "non-pacific"] = typer.Option(..., help="Region tile is in."),
     year: str = typer.Option(..., help="Year to count scenes for."),
     include_t2: bool = typer.Option(
@@ -59,8 +56,6 @@ def count_scenes(
     ),
 ) -> int:
     """Count (tier 1 or all) scenes per tile and year."""
-
-    tile_id_tuple = tuple(map(int, tile_id.split("_")))
     grid = get_gridspec(region=region)
     geobox = grid.tile_geobox(tile_id_tuple)
 
@@ -91,7 +86,7 @@ def run(
     year: Annotated[str, typer.Option()],
     version: Annotated[str, typer.Option()],
     region: Annotated[Literal["pacific", "non-pacific"], typer.Option()],
-    bucket: Annotated[str, typer.Option(help="S3 bucket for data.")] = BUCKET,
+    bucket: Annotated[str | None, typer.Option(help="S3 bucket for data.")] = None,
     owner_pacific: Annotated[
         str,
         typer.Option(help=f"Short owner prefix for Pacific (e.g. '{PACIFIC_OWNER}')."),
@@ -103,6 +98,7 @@ def run(
     product_owner: Annotated[str | None, typer.Option(help="Override the region-derived owner prefix.")] = None,
     overwrite: Annotated[bool, typer.Option()] = False,
     decimated: Annotated[bool, typer.Option()] = False,
+    integration_test: Annotated[bool, typer.Option()] = False,
     mask_shadow: Annotated[
         bool,
         typer.Option(help="True to mask cloud shadows, false to not mask them (leave them in). Defaults to True."),
@@ -118,7 +114,7 @@ def run(
     xy_chunk_size: Annotated[int, typer.Option()] = 2048,
     geomad_threads: Annotated[int, typer.Option()] = 10,
 ) -> None:
-    f"""Run GeoMAD processing on a single tile for a year.
+    """Run GeoMAD processing on a single tile for a year.
 
     Searches USGS STAC for Landsat scenes covering the given tile and year,
     applies cloud masking, computes the geometric median and median absolute
@@ -128,14 +124,21 @@ def run(
     controlled by --ls7-buffer-years is used to gather enough clear
     observations. Pacific tiles may additionally include Tier 2 data.
     """
+    bucket = bucket or get_bucket()  # Default
+
     logger.info(
         f"tile={tile_id} year={year} version={version} region={region} overwrite={overwrite} decimated={decimated} "
-        f"all_bands={all_bands} mask_shadow={mask_shadow} memory={memory_limit} workers={n_workers} "
-        f"threads={threads_per_worker} chunk={xy_chunk_size} geomad_threads={geomad_threads}",
+        f"all_bands={all_bands} mask_shadow={mask_shadow} geomad_threads={geomad_threads}",
+    )
+    logger.info(
+        f"Dask config: n_workers={n_workers}, threads_per_worker={threads_per_worker}, "
+        f"memory_limit={memory_limit}, xy_chunk_size={xy_chunk_size}"
     )
 
     if version != GEOMAD_VERSION:
         logger.info(f"Overriding the latest GeoMAD version ({GEOMAD_VERSION}) with the specified version ({version}).")
+
+    tile_id_tuple = parse_tile_id(tile_id)
 
     year_int = int(year)
     search_year = year
@@ -143,7 +146,7 @@ def run(
 
     min_scenes_threshold = 20
 
-    scene_count_without_t2 = count_scenes(tile_id=tile_id, year=year, region=region, include_t2=False)
+    scene_count_without_t2 = _count_scenes(tile_id_tuple=tile_id_tuple, year=year, region=region, include_t2=False)
     logger.info(f"Scene count (tier 1) for tile/year: {scene_count_without_t2}")
 
     if scene_count_without_t2 >= min_scenes_threshold:
@@ -152,7 +155,7 @@ def run(
 
     else:
         logger.info(f"Scene count ({scene_count_without_t2}) is below {min_scenes_threshold}, trying with T2 too.")
-        scene_count_with_t2 = count_scenes(tile_id=tile_id, year=year, region=region, include_t2=True)
+        scene_count_with_t2 = _count_scenes(tile_id_tuple=tile_id_tuple, year=year, region=region, include_t2=True)
 
         if scene_count_with_t2 >= min_scenes_threshold:
             logger.info(f"Scene count with T2 ({scene_count_with_t2}) is sufficient, using T1 and T2.")
@@ -169,12 +172,9 @@ def run(
                 year_start = year_int - ls7_buffer_years
                 year_end = year_int + ls7_buffer_years
                 search_year = f"{year_start}/{year_end}"
-                typer.echo(f"Using {ls7_buffer_years}-year buffered temporal search for LS7 era: {search_year}")
+                logger.info(f"Using {ls7_buffer_years}-year buffered temporal search for LS7 era: {search_year}")
             else:
                 logger.info("Not in LS7 era so not using buffered temporal search.")
-
-    # Set up variables and check
-    tile_id_tuple = tuple(map(int, tile_id.split("_")))
 
     grid = get_gridspec(region=region)
     geobox = grid.tile_geobox(tile_id_tuple)
@@ -182,58 +182,44 @@ def run(
     # Resolve prefix based on tile region and owner override
     owner = owner_for_region(region, owner_pacific, owner_non_pacific, product_owner)
 
-    # TODO: Handle different bucket formats more robustly. For now we support:
-    # "data.ldn.auspatious.com" to "https://data.ldn.auspatious.com"
-    # "dep-public-staging" to "https://dep-public-staging.s3.us-west-2.amazonaws.com"
-    # "https://data.source.coop" to "https://data.source.coop"
-    if SOURCE_COOP_PUBLIC_URL:
-        full_path_prefix = SOURCE_COOP_PUBLIC_URL
-    elif bucket.startswith("https://"):
-        full_path_prefix = bucket
-    elif "." in bucket:
-        full_path_prefix = f"https://{bucket}"
-    else:
-        full_path_prefix = f"https://{bucket}.s3.{AWS_REGION}.amazonaws.com"
+    full_path_prefix = get_full_path_prefix(bucket)
+    logger.info(f"Full path prefix: {full_path_prefix}")
 
     if decimated:
-        typer.echo("Warning, using decimated (low resolution) for testing purposes.")
-        geobox = geobox.zoom_out(10)  # TODO: Reenable.
-        # geobox = geobox.zoom_out(100) # For faster testing.
+        logger.warning("Warning, using decimated (low resolution) for testing purposes.")
+        geobox = geobox.zoom_out(10)
+
+    if integration_test:
+        logger.warning(
+            "Integration test mode: using 5x5 pixel geobox and limiting to 3 items for very fast processing."
+        )
+        geobox = geobox[0:5, 0:5]
+        search_kwargs["max_items"] = 3
+        n_workers = 1
+        threads_per_worker = 1
+        memory_limit = "1GB"
+        # TODO: Implement these to further speed up.
+        # mask_clouds_kwargs["filters"] = None
+        # geomad_options["maxiters"] = 1
 
     # Configure for dask and reading data
     _ = configure_s3_access(requester_pays=True)
-    # Configure for checking item existence
-    s3_client = boto3.client("s3")  # Only needed for non-Source.Coop.
 
-    # Check if we've done this tile before
-    itempath = PrefixedS3ItemPath(
-        key_prefix=SOURCE_COOP_PREFIX_GEOMAD if SOURCE_COOP_PUBLIC_URL else None,
-        prefix=owner,
-        bucket=bucket,
-        sensor=SENSOR,
-        dataset_id=GEOMAD_DATASET_ID,
-        version=version,
-        time=year,
-        full_path_prefix=full_path_prefix,  # public URL for STAC hrefs + rasterio reads
+    _, prefix_geomad, _ = get_source_coop_config()
+
+    components = build_pipeline_components(
+        tile_id_tuple,
+        year,
+        version,
+        bucket,
+        owner,
+        GEOMAD_DATASET_ID,
+        prefix_geomad if is_source_coop() else None,
+        overwrite,
     )
-    stac_document = itempath.stac_path(tile_id_tuple, absolute=True)
-    stac_key = itempath.stac_path(tile_id_tuple, absolute=False)
-
-    write_session = get_write_session()
-    write_client = get_write_client(write_session)
-
-    aws_client_to_use = write_client if SOURCE_COOP_PREFIX_GEOMAD else s3_client
-
-    # If we don't want to overwrite, and the destination file already exists, skip it
-    # Use the write client to check if the item already exists at the destination, since it may have different creds.
-    if not overwrite and object_exists(bucket, stac_key, client=aws_client_to_use):
-        typer.echo(f"Item already exists at {stac_document}, skipping.")
-        return
-    else:
-        if not overwrite:
-            typer.echo(f"Item does not exist at {stac_document}, processing tile.")
-
-    load_kwargs = {}
+    if components is None:
+        return  # Task exists and overwrite is False, so skipping processing.
+    itempath, write_client, stac_creator, writer, stac_writer = components
 
     # Searcher finds STAC Items
     searcher = PystacSearcher(
@@ -261,27 +247,9 @@ def run(
             "qa_pixel": "ldn.geomad.fuse_qa_pixel",
         },
         fail_on_error=False,  # We don't control the Landsat data so it may have issues. We load what we can.
-        **load_kwargs,
     )
 
-    writer = AwsDsCogWriter(
-        itempath,
-        write_multithreaded=True,
-        write_function=make_write_function(write_session),
-    )
-
-    stac_writer = AwsStacWriter(
-        itempath,
-        client=aws_client_to_use,
-    )
-
-    # Metadata creator
-    stac_creator = StacCreator(
-        collection_url_root=f"{full_path_prefix}/#{owner}_{SENSOR}_{GEOMAD_DATASET_ID}/",
-        itempath=itempath,
-        with_raster=True,
-    )
-
+    # TODO: Make count band use 255 as nodata, rather than 0.
     processor = GeoMADProcessor(
         geomad_options=dict(
             work_chunks=(100, 100),
@@ -314,28 +282,27 @@ def run(
                 searcher=searcher,
                 loader=loader,
                 processor=processor,
+                logger=logger,
                 writer=writer,
                 stac_creator=stac_creator,
                 stac_writer=stac_writer,
             ).run()
-            typer.echo(f"Wrote {len(paths)} files...")
+            logger.info(
+                f"Completed processing. Wrote {len(paths)} files to {itempath.stac_path(tile_id_tuple, absolute=True)}"
+            )
 
     except EmptyCollectionError:
-        typer.echo("No items found for this tile")
+        logger.exception("No items found for this tile")
         sys.exit(EXIT_SKIP)
 
     except InsufficientScenesError as e:
-        typer.echo(f"Failed to process with error: {e}")
+        logger.exception(f"Failed to process with error: {e}")
         sys.exit(EXIT_SKIP)
 
     except KilledWorker as e:
-        typer.echo(f"Failed to process with error: {e}")
+        logger.exception(f"Failed to process with error: {e}")
         sys.exit(EXIT_OOM)
 
-    except Exception as e:
-        typer.echo(f"Failed to process with error: {e}")
+    except Exception:
+        logger.exception("Failed to process with error")
         raise  # let it exit 1 naturally with full traceback
-
-    typer.echo(f"Finished writing to {stac_document}")
-
-    return
