+++ ldn/cli_lulc.py
@@ -3,16 +3,16 @@ from typing import Annotated, Literal
 
 import typer
 
-from ldn.classify import run_classify_task
+from ldn.lulc import run_classify_task
 from ldn.utils import (
     AWS_REGION,
-    BUCKET,
     GEOMAD_VERSION,
+    LULC_VERSION,
     MODEL_VERSION,
     NON_PACIFIC_OWNER,
     PACIFIC_OWNER,
-    PREDICTION_VERSION,
     LdnError,
+    get_bucket,
     owner_for_region,
 )
 
@@ -22,33 +22,37 @@ logger = logging.getLogger(__name__)
 
 @classify_app.command()
 def run(
-    tile_id: str = typer.Option(..., help="Tile ID to predict LULC for."),
-    year: str = typer.Option(..., help="Year to predict LULC for."),
+    tile_id: str = typer.Option(..., help="Tile ID to classify LULC for."),
+    year: str = typer.Option(..., help="Year to classify LULC for."),
     version: str = typer.Option(
-        PREDICTION_VERSION,
-        help=f"Version of the model to use e.g. '{PREDICTION_VERSION}'.",
+        LULC_VERSION,
+        help=f"Version of training data to output e.g. '{LULC_VERSION}'.",
     ),
     version_geomad: str = typer.Option(
         GEOMAD_VERSION,
         help=f"Version of the GeoMAD data to use e.g. '{GEOMAD_VERSION}'.",
     ),
     region: Literal["pacific", "non-pacific"] = typer.Option(
-        ..., help="Region to predict LULC for. Can be 'pacific' or 'non-pacific'."
+        ..., help="Region tile belongs to. Can be 'pacific' or 'non-pacific'."
     ),
-    bucket: str = typer.Option(BUCKET, help="S3 bucket for data."),
+    bucket: Annotated[str | None, typer.Option(help="S3 bucket for data.")] = None,
     owner_pacific: str = typer.Option(PACIFIC_OWNER, help="S3 owner prefix for Pacific data."),
     owner_non_pacific: str = typer.Option(NON_PACIFIC_OWNER, help="S3 owner prefix for non-Pacific data."),
     product_owner: str | None = typer.Option(None, help="Override the region-derived owner prefix."),
     model_path: str = typer.Option(
         # TODO: defaults to pacific. Later have per region/time period models.
         f"https://s3.{AWS_REGION}.amazonaws.com/data.ldn.auspatious.com/models/{MODEL_VERSION}/pacific/2020/lulc_random_forest_model_pacific_2020.joblib",
-        help="Model to use for prediction.",
+        help="Model to use for LULC classification.",
     ),
     decimated: bool = typer.Option(
         False,
-        help="Whether to use decimated data for prediction. Decimated data is faster to predict but less accurate.",
+        help="Lower resolution data for faster processing/testing.",
     ),
-    overwrite: bool = typer.Option(False, help="Whether to overwrite existing prediction."),
+    integration_test: bool = typer.Option(
+        False,
+        help="Integration test mode: use decimated data for faster processing.",
+    ),
+    overwrite: bool = typer.Option(False, help="Whether to overwrite existing LULC classification."),
     probability_threshold: float = typer.Option(
         30.0,
         help="Probability threshold (0-100) for classifying a pixel as the target class. "
@@ -71,6 +75,7 @@ def run(
     if int(year) < 2000 or int(year) > 2025:
         raise LdnError("Year must be between 2000 and 2025.")
 
+    bucket = bucket or get_bucket()  # Default
     owner = owner_for_region(region, owner_pacific, owner_non_pacific, product_owner)
 
     run_classify_task(
@@ -84,6 +89,7 @@ def run(
         model_path=model_path,
         xy_chunk_size=xy_chunk_size,
         decimated=decimated,
+        integration_test=integration_test,
         overwrite=overwrite,
         probability_threshold=probability_threshold,
         nodata_value=nodata_value,
