+++ ldn/utils.py
@@ -1,9 +1,46 @@
 import logging
+import os
+import re
 from typing import Literal
 
 from dep_tools.grids import COUNTRIES_AND_CODES as DEP_COUNTRIES_AND_CODES
+from dotenv import load_dotenv
 
 logger = logging.getLogger(__name__)
+
+
+load_dotenv()
+
+
+def get_bucket() -> str:
+    """Return the bucket name from the environment variable."""
+    bucket = os.environ.get("BUCKET")
+    if not bucket:
+        raise LdnError("BUCKET environment variable must be set.")
+    return bucket
+
+
+def get_source_coop_config() -> tuple[str | None, str | None, str | None]:
+    """Return (public_url, prefix_geomad, prefix_lulc) from environment."""
+    return (
+        os.environ.get("SOURCE_COOP_PUBLIC_URL") or None,
+        os.environ.get("SOURCE_COOP_PREFIX_GEOMAD") or None,
+        os.environ.get("SOURCE_COOP_PREFIX_LULC") or None,
+    )
+
+
+# This is a function instead of a module-level variable to ensure it reflects any changes
+# to the environment variables during testing (e.g. via monkeypatch).
+def is_source_coop() -> bool:
+    """Return True if all Source.Coop environment variables are set."""
+    url, prefix_geomad, prefix_lulc = get_source_coop_config()
+    return bool(url) and bool(prefix_geomad) and bool(prefix_lulc)
+
+
+# Our custom exception class for the project. Good for filtering errors in processing.
+class LdnError(Exception):
+    """Base exception for the ldn-lulc project."""
+
 
 SIDS_COUNTRIES_AND_CODES = {
     # Caribbean
@@ -74,70 +111,23 @@ ALL_COUNTRIES = {**SIDS_COUNTRIES_AND_CODES, **DEP_COUNTRIES_AND_CODES}
 # Get SIDS countries that are not in DEP for CI Grid use.
 NON_DEP_COUNTRIES = {k: v for k, v in SIDS_COUNTRIES_AND_CODES.items() if k not in DEP_COUNTRIES_AND_CODES}
 
-# These tiles are representative of different environments e.g. forest, atoll, volcanic, elevated, urban, beach,
-# wetland, grassland, cropland, etc, Give me more if more than 5 are needed
-PACIFIC_TRAINING_TILES = [
-    # Papua New Guinea: Dense tropical rainforest & highland montane forest.
-    ("028_030", "pacific", {"Papua New Guinea": "PNG"}),  # Capital city and coast.
-    ("024_034", "pacific", {"Papua New Guinea": "PNG"}),  # Highland.
-    ("023_031", "pacific", {"Papua New Guinea": "PNG"}),  # River delta.
-    # Kiribati: Low-lying coral atoll, almost entirely at sea level, classic open-ocean/lagoon environment.
-    ("058_043", "pacific", {"Kiribati": "KIR"}),
-    ("059_040", "pacific", {"Kiribati": "KIR"}),
-    # Vanuatu: Active volcanic islands with crater lakes, lava fields, and cloud forest.
-    ("051_023", "pacific", {"Vanuatu": "VUT"}),
-    ("053_018", "pacific", {"Vanuatu": "VUT"}),  # Mt Yasur volcano.
-    ("052_022", "pacific", {"Vanuatu": "VUT"}),  # Lava lake.
-    # Samoa: Elevated volcanic interior with waterfalls and lava tubes, fringed by reef/beach coastline.
-    # 2 tiles pretty much covers all of Samoa.
-    ("074_025", "pacific", {"Samoa": "WSM"}),
-    ("075_025", "pacific", {"Samoa": "WSM"}),
-    # Fiji: The most "mixed urban + agricultural" of the group, with sugarcane croplands,
-    # mangrove wetlands, and a developed capital (Suva)
-    ("063_020", "pacific", {"Fiji": "FJI"}),  # Elevation.
-    ("066_022", "pacific", {"Fiji": "FJI"}),  # AM-crossing.
-    ("064_020", "pacific", {"Fiji": "FJI"}),  # Suva urban area.
-    # Palau for raised limestone/rock island jungle
-    ("013_050", "pacific", {"Palau": "PLW"}),
-    # New Caledonia for maquis shrubland / lagoon
-    ("050_015", "pacific", {"New Caledonia": "NCL"}),
-]
-
-
-GEOMAD_VERSION = "0-2-1"
-PREDICTION_VERSION = "0-0-4"
-MODEL_VERSION = "0-0-4"
-TRAINING_DATA_VERSION = "0-0-4"
-
-# aws s3 cp test.txt s3://us-west-2.opendata.source.coop/auspatious/geomad-sids/test.txt
-# aws s3 rm s3://us-west-2.opendata.source.coop/auspatious/geomad-sids/test.txt
-# aws s3 rm s3://us-west-2.opendata.source.coop/auspatious/lulc-sids/dep_ls_lulc_prediction/0-0-4-test/ --recursive
-SOURCE_COOP_PUBLIC_URL = "https://data.source.coop"  # public read URL for STAC hrefs
-SOURCE_COOP_PREFIX_GEOMAD = "auspatious/geomad-sids"  # For source.coop.
-SOURCE_COOP_PREFIX_PREDICTION = "auspatious/lulc-sids"  # For source.coop.
-# SOURCE_COOP_PUBLIC_URL = None # For non-Source.Coop buckets.
-# SOURCE_COOP_PREFIX_GEOMAD = None
-# SOURCE_COOP_PREFIX_PREDICTION = None
-
-# BUCKET = "data.ldn.auspatious.com"
-# BUCKET = "dep-public-staging"
-BUCKET = "us-west-2.opendata.source.coop"
-
+GEOMAD_VERSION = "0-3-0"
+LULC_VERSION = "0-1-0"
+MODEL_VERSION = "0-1-0"
+TRAINING_DATA_VERSION = "0-1-0"
 
 PACIFIC_OWNER = "dep"
 NON_PACIFIC_OWNER = "ci"
 
 SENSOR = "ls"
 GEOMAD_DATASET_ID = "geomad"
-PREDICTION_DATASET_ID = "lulc_prediction"
+LULC_DATASET_ID = "lulc"
 AWS_REGION = "us-west-2"
 
 LS7_YEAR_THRESHOLD = 2012
-training_data_year = "2020"
-
-class_attr = "lulc"
-
-wgs84 = "EPSG:4326"
+TRAINING_DATA_YEAR = "2020"
+CLASS_ATTR = "lulc"  # For training data creation.
+WGS84 = "EPSG:4326"
 
 
 def owner_for_region(
@@ -155,74 +145,215 @@ def owner_for_region(
     return owner_pacific if region == "pacific" else owner_non_pacific
 
 
-def dataset_prefix(owner: str, dataset_id: str) -> str:
-    """Build the full dataset prefix from owner and dataset_id.
+def dataset_prefix(owner: str | None, dataset_id: str) -> str:
+    """Build the full dataset prefix from owner, sensor, and dataset_id (owner optional).
 
     Args:
         owner: Short owner prefix (e.g. "dep" or "ci").
-        dataset_id: Dataset identifier (e.g. "geomad" or "lulc_prediction").
+        dataset_id: Dataset identifier (e.g. "geomad" or "lulc").
 
     Returns:
-        Full prefix like "dep_ls_geomad" or "ci_ls_lulc_prediction".
+        Full prefix like "dep_ls_geomad" or "ci_ls_lulc" or "ls_geomad" if owner is None.
     """
-    return f"{owner}_{SENSOR}_{dataset_id}"
+    if owner:
+        return f"{owner}_{SENSOR}_{dataset_id}"
+    else:
+        return f"{SENSOR}_{dataset_id}"
 
 
-def get_geomad_stac_geoparquet_url(
-    region: Literal["pacific", "non-pacific"],
-    product_owner: str | None = None,
-    version: str | None = None,
+def get_stac_geoparquet_key(
+    dataset_id: str,
+    version: str,
+    source_coop_prefix: str | None = None,
 ) -> str:
-    """Build the STAC-Geoparquet URL for GeoMAD data in a given region.
+    """Return the S3 key for the STAC-Geoparquet file (no bucket, no scheme).
 
     Args:
-        region: Either "pacific" or "non-pacific".
-        product_owner: Optional override for the region-derived owner prefix.
-        version: GeoMAD version string. Defaults to GEOMAD_VERSION.
+        dataset_id: The dataset ID (e.g. 'geomad').
+        version: Version string.
+        source_coop_prefix: Source.Coop prefix if applicable, else None.
 
     Returns:
-        HTTPS URL to the STAC-Geoparquet file.
+        S3 key string e.g. 'auspatious/geomad-sids/ls_geomad/0-2-1/ls_geomad.parquet'
+        or 'ls_geomad/0-2-1/ls_geomad.parquet' for a standard bucket.
     """
-    ver = version if version is not None else GEOMAD_VERSION
-    bucket = BUCKET  # TODO: Make this a param?
-    owner = owner_for_region(region, product_owner=product_owner)
-    prefix = dataset_prefix(owner, GEOMAD_DATASET_ID)
-    if SOURCE_COOP_PUBLIC_URL:
-        return f"{SOURCE_COOP_PUBLIC_URL}/{SOURCE_COOP_PREFIX_GEOMAD}/{prefix}/{ver}/{prefix}.parquet"
-    return f"https://s3.{AWS_REGION}.amazonaws.com/{bucket}/{prefix}/{ver}/{prefix}.parquet"
+    combined_short = dataset_prefix(None, dataset_id)
+    key = f"{combined_short}/{version}/{combined_short}.parquet"
+    if source_coop_prefix:
+        return f"{source_coop_prefix}/{key}"
+    return key
 
 
-def get_geomad_item_id(
-    region: Literal["pacific", "non-pacific"],
-    tile_id: str,
-    year: str,
-    product_owner: str | None = None,
-) -> str:
-    """Build the STAC item ID for a GeoMAD tile.
+def get_geomad_stac_geoparquet_url(bucket: str, version: str) -> str:
+    """Return the URL to the GeoMAD STAC-Geoparquet file for use with rustac/DuckDB.
+
+    For Source.Coop, returns a public HTTPS URL (publicly readable).
+    For all other buckets, returns a regional S3 HTTPS URL since DuckDB needs
+    a resolvable endpoint and cannot handle custom-domain buckets directly.
 
     Args:
-        region: Either "pacific" or "non-pacific".
-        tile_id: Grid tile identifier (e.g. "058_043").
-        year: Year string (e.g. "2020").
-        product_owner: Optional override for the region-derived owner prefix.
+        bucket: The S3 bucket name or custom domain.
+        version: GeoMAD version string.
 
     Returns:
-        The full STAC item ID string.
+        URL to the STAC-Geoparquet file.
     """
-    owner = owner_for_region(region, product_owner=product_owner)
-    prefix = dataset_prefix(owner, GEOMAD_DATASET_ID)
-    return f"{prefix}_{tile_id}_{year}"
+    source_coop_url, prefix_geomad, _ = get_source_coop_config()
+    key = get_stac_geoparquet_key(GEOMAD_DATASET_ID, version, prefix_geomad)
+    if source_coop_url:
+        return f"{source_coop_url}/{key}"
+    return f"https://s3.{AWS_REGION}.amazonaws.com/{bucket}/{key}"
 
 
 def get_analysis_epsg(
     region: Literal["pacific", "non-pacific"],
 ) -> Literal["EPSG:3832", "EPSG:6933"]:
+    """Return the appropriate EPSG code for analysis based on the region."""
     if region == "pacific":
         return "EPSG:3832"
+    return "EPSG:6933"
+
+
+def parse_years(years: str) -> list[int]:
+    """Parse a years string into a list of integers.
+
+    Accepts a comma-separated list (e.g. '2020,2021') or a range (e.g. '2010-2023').
+    Ranges must be ascending (e.g. '2020-2023').
+    """
+    if "," in years:
+        return [int(y.strip()) for y in years.split(",")]
+    elif "-" in years:
+        start_year, end_year = map(int, years.split("-"))
+        if start_year > end_year:
+            raise ValueError(f"Invalid year range: '{years}'. Start year must be <= end year.")
+        return list(range(start_year, end_year + 1))
     else:
-        return "EPSG:6933"
+        return [int(years)]
 
 
-# Our custom exception class for the project. Good for filtering errors in processing.
-class LdnError(Exception):
-    """Base exception for the ldn-lulc project."""
+# TODO: _source_coop_prefix() returns source_coop_prefix
+def resolve_dataset(
+    dataset: Literal["geomad", "lulc"],
+    version_geomad: str,
+    version_lulc: str,
+) -> tuple[str, str, str | None]:
+    """Return (dataset_id, version, source_coop_prefix) for the given dataset name."""
+    _, prefix_geomad, prefix_lulc = get_source_coop_config()
+    if dataset == "geomad":
+        return GEOMAD_DATASET_ID, version_geomad, prefix_geomad
+    return LULC_DATASET_ID, version_lulc, prefix_lulc
+
+
+def parse_tile_id(tile_id: str) -> tuple[int, int]:
+    """Parse a tile ID string into a tuple of two integers.
+
+    Accepts any of '_', ',', or '-' as separators, e.g. '028_030',
+    '28,30', or '28-30'.
+
+    Args:
+        tile_id: A string containing two integers separated by '_', ',', or '-'.
+
+    Returns:
+        A tuple of two integers (x, y).
+
+    Raises:
+        LdnError: If the tile ID does not split into exactly two integers.
+    """
+    parts = [int(i) for i in re.split(r"[,\-_]", tile_id)]
+    if len(parts) != 2:
+        raise LdnError(f"Tile ID must split into 2 integers, got {parts} from '{tile_id}'")
+    return parts[0], parts[1]
+
+
+def get_public_https_prefix(bucket: str) -> str:
+    """Return the public HTTPS URL prefix for a given bucket.
+
+    Used for STAC hrefs and other public HTTP contexts (e.g. geoparquet URLs).
+
+    Args:
+        bucket: The S3 bucket name or custom domain.
+
+    Returns:
+        A public HTTPS URL prefix string.
+    """
+    source_coop_url, _, _ = get_source_coop_config()
+    if source_coop_url:
+        return source_coop_url
+    if "." in bucket:
+        return f"https://{bucket}"
+    return f"https://s3.{AWS_REGION}.amazonaws.com/{bucket}"
+
+
+def get_collection_url_root(
+    bucket: str,
+    owner: str,
+    sensor: str,
+    dataset_id: str,
+) -> str:
+    """Return the collection URL root for STAC metadata.
+
+    Handles three bucket styles:
+    - Source.Coop (e.g. 'us-west-2.opendata.source.coop'): uses the public Source.Coop HTTPS URL.
+    - Custom-domain bucket (e.g. 'data.ldn.auspatious.com'): uses 'https://{bucket}'.
+    - Standard S3 bucket (e.g. 'dep-public-staging'): uses the regional S3 endpoint URL.
+
+    Args:
+        bucket: The S3 bucket name or custom domain.
+        owner: The owner prefix (e.g. 'dep', 'ci').
+        sensor: The sensor string (e.g. 'ls').
+        dataset_id: The dataset ID (e.g. 'geomad').
+
+    Returns:
+        A public HTTPS URL string suitable for use as a STAC collection URL root.
+    """
+    base = get_public_https_prefix(bucket)
+    return f"{base}/#{owner}_{sensor}_{dataset_id}/"
+
+
+def get_full_path_prefix(bucket: str) -> str:
+    """Return the path prefix rasterio should use to read back written files.
+
+    Handles three bucket styles:
+    - Source.Coop (e.g. 'us-west-2.opendata.source.coop'): returns the public
+      Source.Coop HTTPS URL since files are publicly readable without auth.
+    - Custom-domain bucket (e.g. 'data.ldn.auspatious.com'): returns 's3://{bucket}'
+      so rasterio uses boto3 credentials to authenticate.
+    - Standard S3 bucket (e.g. 'dep-public-staging'): returns 's3://{bucket}'
+      so rasterio uses boto3 credentials to authenticate.
+
+    Args:
+        bucket: The S3 bucket name or custom domain.
+
+    Returns:
+        A URL prefix string suitable for rasterio to open files.
+    """
+    source_coop_url, _, _ = get_source_coop_config()
+    if source_coop_url:
+        return source_coop_url
+    return f"s3://{bucket}"
+
+
+def get_s3_mosaic_write_path(
+    bucket: str,
+    dataset_id: str,
+    version: str,
+    source_coop_prefix: str | None = None,
+) -> str:
+    """Return the S3 write path prefix for a given dataset.
+
+    Handles Source.Coop (includes prefix) and standard/custom-domain buckets.
+
+    Args:
+        bucket: The S3 bucket name or custom domain.
+        dataset_id: The dataset ID (e.g. 'geomad').
+        version: Version string.
+        source_coop_prefix: Source.Coop prefix if applicable, else None.
+
+    Returns:
+        S3 path string e.g. 's3://us-west-2.opendata.source.coop/auspatious/geomad-sids/ls_geomad/0-2-1'
+        or 's3://data.ldn.auspatious.com/ls_geomad/0-2-1'
+    """
+    combined_short = dataset_prefix(None, dataset_id)
+    if source_coop_prefix:
+        return f"s3://{bucket}/{source_coop_prefix}/{combined_short}/{version}"
+    return f"s3://{bucket}/{combined_short}/{version}"
